Created unit tests for creating a connection pool and getting connections
Used the pool builder to set the max number of connections. This is
helpful for the unit tests to be able to take all of the connections.
Created a helper function for getting a connection. Wrote complete unit
tests for opening a connection pool, taking an available connection, and
attempting to take a connection when none are available. Tests for the
`create_tables` and `add_player` functions are stubbed.