Created a separate `Connection` struct that implements database operations
Instead of using the pool to perform database operations, the pool is
used to get a `Connection` that can be used to perform those operations.
Created new modules to separate the pool and connection code. Tests will
need to be rethought as an `add_players` test would be testing
`Connection`. The current tests are also technically integration tests
since they integrate with a running SQLite database. To my knowledge
diesel has no mock database that can be used to run tests against, so
`Connection` and `ConnectionPool` cannot be made generic to accept a
mock that can be used for proper *unit* testing. As such, these tests
should be moved to a top-level tests/ directory.