Overhauled connection and pool errors
Created `PoolCreationError` to represent issues when creating a
connection pool. There is an assertion that query errors cannot happen
when initializing connections to the database. Created a `PoolTimeout`
unit struct that can be returned as an error when getting a connection
from the pool. There is an assertion that query and connection errors
cannot happen here as all connections are already established.