Added timestamps to the matches and names tables
Added created column to the matches table to store a timestamp for that
match. Added last_used column to the names table to store a timestamp
for when a username was last used by a player. Changed the
`add_player_name` function to `add_or_update_player_name` that will
update the last_used timestamp if the (player_id, name) key already
exists in the names table.