Moved database operations under a trait
Moved all database operations under the `DatabaseConnection` trait. This
allows me to provide a blanket implementation for all types that are
`Send` and implement `AsyncConnection`. This means that the same type
can be used in "raw" diesel queries as well as call the custom database
operations. The motivation was to be able leverage transactions to
atomically add groups of related data to the tables at once. This
requires that the type used to create the type can call the custom
database operations.