Initial web server in axum
Created the initial structure for the eventual puck-chasing web server
written using axum. It currently does not serve any frontend and has a
"Hello world" handler at the root. This will be changed when this branch
is merged. For the time being the goal is to build custom types that can
be serialized and sent as JSON to the client and then use the existing
database operations to get the info necessary to build those types.