Added route to get all player names
Added a GET route ("/api/players") that returns a map of each player's
slap ID to a list of their known usernames. The results of the database
operations were mapped into the final return type, returning an internal
server error if anything failed (this may be subject to change in the
future). This is likely how other routes will end up looking as well.
I'm still unsure if custom types that implement `Serialize` will be
needed or if I can leverage the `json!` macro for all of my JSON
returning needs.