Created a macro that can generate string types
Created a macro that reduces the amount of duplicated code for string
types. Instead of implementing the same thing again and again for these
types, new implementations can be added to the macro that then applies
for all types generated with it. A similar macro will be created for the
integer types.